dashboard folder an route.js file deleted from api folder; content of route.js moved to data.ts Co-authored-by: jgmsv <jose.guilherme.vieira@gmail.com>
