+++ src/app/lib/data.ts
@@ -8,7 +8,6 @@ import { unstable_noStore as noStore } from 'next/cache';
 
 
 
-
 export async function fetchAllCourses(): Promise<number> {
     noStore();
     try {
@@ -126,5 +125,28 @@ export async function fetchCoursesByLocation(
 }
 
 
+export async function fetchPersonByRole(): Promise<string> {
+  noStore();
+
+    try {
+
+      //get login info => person id
+      //const personId = fetchPersonById();
+      //filter by role
+      // side bar adjustments depending on the person's role
+
+      const response = await fetch('http://localhost:8080/api/v1/persons/role/:role');
+      if(!response.ok) {
+        throw new Error('Failed to fetch person by role.');
+      }
+
+      const person: Person = await response.json();
+
+      return person.role;
+    } catch (error) {
+      console.error('Database error:', error);
+        throw new Error('Failed to fetch person by role.');
+    }
+}
